Move updater to utils
